More forgiving error states. MessageBlock refctor.
